requirements.txt
# ═══════════════════════════════════════════════════════════════
# Enterprise GenAI Platform — Python Dependencies
# Pin major versions for reproducibility; use ~= for minor flexibility
# ═══════════════════════════════════════════════════════════════

# ─── Core API Framework ─────────────────────────────────────────
fastapi~=0.111.0
uvicorn[standard]~=0.29.0
pydantic~=2.7.0
pydantic-settings~=2.2.0

# ─── LLM & Agent Orchestration ──────────────────────────────────
openai~=1.30.0
anthropic~=0.26.0
langchain~=0.2.0
langchain-openai~=0.1.0
langchain-community~=0.2.0
langgraph~=0.1.0

# ─── Embeddings ─────────────────────────────────────────────────
sentence-transformers~=3.0.0
# Optional: for reranking
# cross-encoder available via sentence-transformers

# ─── Vector Store ───────────────────────────────────────────────
faiss-cpu~=1.8.0            # Use faiss-gpu on GPU instances

# ─── Document Processing ────────────────────────────────────────
pymupdf~=1.24.0              # PDF parsing (PyMuPDF / fitz)
python-docx~=1.1.0           # DOCX parsing
# Optional OCR
# pytesseract~=0.3.10
# Pillow~=10.3.0

# ─── Database ───────────────────────────────────────────────────
duckdb~=0.10.0               # Analytical SQL queries
asyncpg~=0.29.0              # Async PostgreSQL driver
sqlalchemy~=2.0.0
psycopg2-binary~=2.9.9       # Sync PostgreSQL (for MLflow)

# ─── Caching ────────────────────────────────────────────────────
redis~=5.0.0
hiredis~=2.3.0               # C extension for Redis performance

# ─── MLOps ──────────────────────────────────────────────────────
mlflow~=2.13.0

# ─── Hybrid Search ──────────────────────────────────────────────
rank-bm25~=0.2.2             # BM25 sparse retrieval

# ─── HTTP Client ────────────────────────────────────────────────
httpx~=0.27.0

# ─── Data & Numerics ────────────────────────────────────────────
numpy~=1.26.0
pandas~=2.2.0
tabulate~=0.9.0              # Markdown table output for SQL results

# ─── Dashboard ──────────────────────────────────────────────────
streamlit~=1.35.0
plotly~=5.22.0

# ─── Observability ──────────────────────────────────────────────
# prometheus-client~=0.20.0  # Uncomment for native Prometheus client
# opentelemetry-sdk~=1.24.0  # Uncomment for OTLP export

# ─── Development & Testing ──────────────────────────────────────
pytest~=8.2.0
pytest-asyncio~=0.23.0
pytest-cov~=5.0.0
httpx~=0.27.0                # Test client for FastAPI
python-multipart~=0.0.9     # File upload support in FastAPI
